Changed u8 back to u32

+++ tests/ui/arch/atomics/atomic_compare_exchange.rs
@@ -12,9 +12,9 @@ pub fn main(#[spirv(descriptor_set = 0, binding = 0, storage_buffer)] buffer: &m
     unsafe {
         let old = spirv_std::arch::atomic_compare_exchange::<
             _,
-            { Scope::CrossDevice as u8 },
-            { Semantics::UNIFORM_MEMORY.bits() as u8 },
-            { Semantics::UNIFORM_MEMORY.bits() as u8 },
+            { Scope::CrossDevice as u32 },
+            { Semantics::UNIFORM_MEMORY.bits() as u32 },
+            { Semantics::UNIFORM_MEMORY.bits() as u32 },
         >(&mut *buffer.index_unchecked_mut(0), 5, 10);
     }
 }
